Enable advanced helloworld features and update build process
This commit introduces advanced helloworld features and adds the `--enable-helloworld-advanced` flag for configuration. It modifies the build and install process in `build.sh` and adds advanced features support in `config.m4`. It also introduces a structure in `php_helloworld.h` for the global greeting variable. The `helloworld.php` script has been updated to include an advanced helloworld function.