Add run configuration file for CLion
A new run configuration file for the CLion IDE has been created. This file is necessary for running the project in CLion and includes parameters for the 'helloworld' module and 'smoke.php' file. The configuration includes the specifics of the run environment setup, including directory, input redirection, terminal emulation settings, and more within the .idea/runConfigurations directory.